Rearange to create clarity of purpose
